+++ CHANGELOG.rst
@@ -56,6 +56,7 @@ released 2017-03-30
 =====
 released 2017-03-06
 
+* CHANGED ``--verbose`` flag to ``--verbosity``, allowing finer granularity.
 * FIX `list` (and commands based on it like `calendar`, `at`, and `search`)
   crashed if `--notstarted` was given and allday events were found (introduced
   in 0.9.2)